renamed the service, and file-scoped namespaces
