+++ src/gei/Program.cs
@@ -40,7 +40,7 @@ namespace OctoshiftCLI.GithubEnterpriseImporter
                 .AddSingleton<ISourceGithubApiFactory>(sp => sp.GetRequiredService<GithubApiFactory>())
                 .AddSingleton<SecretScanningAlertServiceFactory>()
                 .AddSingleton<CodeScanningAlertServiceFactory>()
-                .AddSingleton<GhesVersionCheckerService>()
+                .AddSingleton<GhesVersionChecker>()
                 .AddHttpClient("NoSSL")
                 .ConfigurePrimaryHttpMessageHandler(() => new HttpClientHandler()
                 {
